Исправил ошибку добавления товара в корзину

+++ Bookstore.Application/CommandAndQuery/Books/Commands/DeleteBook/DeleteBookCommandHandler.cs
@@ -1,30 +1,35 @@
-﻿using Bookstore.Application.Interfaces;
-using MediatR;
-
-namespace Bookstore.Application.CommandAndQuery.Books.Commands.DeleteBook
+﻿namespace Bookstore.Application.CommandAndQuery.Books.Commands.DeleteBook
 {
     public class DeleteBookCommandHandler : IRequestHandler<DeleteBookCommand, Unit>
     {
         private readonly IBaseDbContext _context;
+        public DeleteBookCommandHandler(IBaseDbContext context) => _context = context;
 
-        public DeleteBookCommandHandler(IBaseDbContext context) =>
-            _context = context;
-
-
-        public async Task<Unit> Handle(DeleteBookCommand request,
-            CancellationToken cancellationToken)
+        public async Task<Unit> Handle(DeleteBookCommand request, CancellationToken cancellationToken)
         {
-            var bookFromDb = await _context.Books.FindAsync(request.Id);
-
-            if (bookFromDb is null)
+            try
             {
-                throw new ArgumentNullException(nameof(bookFromDb));
+                if (request is not null)
+                {
+                    var bookFromDb = await _context.Books.FindAsync(request.Id, cancellationToken);
+
+                    if (bookFromDb is null)
+                    {
+                        throw new ArgumentNullException(nameof(bookFromDb));
+                    }
+                    _context.Books.Remove(bookFromDb);
+                    await _context.SaveChangesAsync(cancellationToken);
+                }
+
+                return Unit.Value;
             }
+            catch(Exception e)
+            {
+                throw e;
+            }
+            
 
-            _context.Books.Remove(bookFromDb);
-            await _context.SaveChangesAsync(cancellationToken);
-
-            return Unit.Value;
+            
         }
     }
 }
