+++ Bookstore.Application/CommandAndQuery/Reviews/Commands/DeleteReview/DeleteReviewCommandHandler.cs
@@ -3,29 +3,28 @@
     public class DeleteReviewCommandHandler : IRequestHandler<DeleteReviewCommand, Unit>
     {
         private readonly IBaseDbContext _context;
-
         public DeleteReviewCommandHandler(IBaseDbContext context) => _context = context;
-        
         public async Task<Unit> Handle(DeleteReviewCommand request, CancellationToken cancellationToken)
         {
             try
             {
-                var reviewFromDb = await _context.Reviews
-                    .FindAsync(request.ReviewId, cancellationToken);
-
-                if (reviewFromDb is null)
+                if (request is not null)
                 {
-                    throw new ArgumentNullException(nameof(Review));
+                    var reviewFromDb = await _context.Reviews
+                        .FindAsync(request.ReviewId, cancellationToken);
+                    if (reviewFromDb is null)
+                    {
+                        throw new ArgumentNullException(nameof(Review));
+                    }
+                    _context.Reviews.Remove(reviewFromDb);
+                    await _context.SaveChangesAsync(cancellationToken);
                 }
-
-                _context.Reviews.Remove(reviewFromDb);
-                await _context.SaveChangesAsync(cancellationToken);
 
                 return Unit.Value;
             }
-            catch (Exception)
+            catch (Exception e)
             {
-                throw;
+                throw e;
             }
         }
     }
